backend/requirements.txt
fastapi~=0.128.0
uvicorn~=0.40.0
sqlalchemy~=2.0.46
psycopg2-binary
pydantic-settings~=2.12.0
python-multipart
websockets
openai~=2.15.0
python-dotenv
requests~=2.32.5
PyPDF2
python-docx